fail

+++ start.sh
@@ -1,2 +1,5 @@
-tmux kill-session -t website
-tmux new -d -s website "cd ~/pub/ && git pull && python3 app.py"
+if tmux has-session -t website; then
+	tmux kill-session -t website;
+fi
+
+tmux new -d -s website "cd ~/pub/ && git pull && python3 app.py";
